requirements.txt
beautifulsoup4
groq
pandas
pdf2image
pdfplumber
pytesseract
python-dotenv
requests
sentence-transformers
streamlit
